requirements.txt
cachelib==0.9.0
click==8.1.3
Flask==2.2.2
Flask-Session==0.4.0
Flask-WTF==1.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mpmath==1.2.1
numpy==1.23.3
opencv-python-headless==4.6.0.66
Pillow==9.3.0
pycryptodomex==3.15.0
sympy==1.11.1
Werkzeug==2.2.2
WTForms==3.0.1
